+++ package.json
@@ -151,7 +151,8 @@
     "vite": "^5.1.4",
     "vite-plugin-environment": "^1.1.3",
     "vite-plugin-eslint": "^1.8.1",
-    "vite-plugin-istanbul": "^5.0.0"
+    "vite-plugin-istanbul": "^5.0.0",
+    "vite-plugin-node-polyfills": "^0.21.0"
   },
   "nyc": {
     "extends": "@istanbuljs/nyc-config-typescript",
